New Test(293237@main): [ Debug ] cookies/cookie-store-register-eventlistener-from-file-protocol.html trips assertion https://bugs.webkit.org/show_bug.cgi?id=292217 rdar://150231572
Reviewed by Jonathan Bedard and Per Arne Vollan.

These newly added Cookie Store tests were failing on WK1.
But the Cookie Store API is not supported on WK1, so we
skip them in the WK1 test expectations.

Also, there is already a LayoutTests/cookies/ folder, so we don't
need to create a LayoutTests/fast/cookie-store folder. This patch
makes a drive-by change to move the test there.

* LayoutTests/cookies/cookie-store-start-listening-for-change-with-null-url-crash-expected.txt: Renamed from LayoutTests/fast/cookie-store/start-listening-for-change-with-null-url-crash-expected.txt.
* LayoutTests/cookies/cookie-store-start-listening-for-change-with-null-url-crash.html: Renamed from LayoutTests/fast/cookie-store/start-listening-for-change-with-null-url-crash.html.
* LayoutTests/platform/mac-wk1/TestExpectations:
* LayoutTests/platform/mac/TestExpectations:

Canonical link: https://commits.webkit.org/294286@main