[text-indent] Unexpected line breaking when text-indent has negative value https://bugs.webkit.org/show_bug.cgi?id=292236
Reviewed by Antti Koivisto.

Decouple min and max width handling with text-indent.

text-indent logic in computeInlinePreferredLogicalWidths assumes that min and max values
of a child box are always the same so that they clear the negative text-indent value
at the same time (the min value is used as base and we just offset the max value using the diff).

* LayoutTests/fast/inline/negative-text-indent-triggers-content-wrap-expected.html: Added.
* LayoutTests/fast/inline/negative-text-indent-triggers-content-wrap.html: Added.
* Source/WebCore/rendering/RenderBlockFlow.cpp:
(WebCore::RenderBlockFlow::computeInlinePreferredLogicalWidths const):

Canonical link: https://commits.webkit.org/294296@main