[WebGPU] fast/webgpu/nocrash/fuzz-275294.html crashes in Debug https://bugs.webkit.org/show_bug.cgi?id=292020 rdar://149971001
Reviewed by Tadeu Zagallo.

When there are no slots in the cache, because for instance a
compute shader running an infinite loop is using 2000 samplers,
we should return nil instead of removing elements from the global
cache.

Removing elements doesn't resolve the lifetime of the MTLSamplerState
object that is cached by the MTLCommandBuffer which is still executing
an infinte loop, for instance.

* LayoutTests/fast/webgpu/nocrash/fuzz-275294-expected.txt:
Update expectations.
* LayoutTests/platform/mac-wk2/TestExpectations:
Set to Pass

* Source/WebGPU/WebGPU/Sampler.mm:
(WebGPU::Sampler::samplerState const):
Correct cache.

Canonical link: https://commits.webkit.org/294328@main