JSToWasm non-IC entrypoints should load the entry point the same way as the IC. https://bugs.webkit.org/show_bug.cgi?id=286946 rdar://143456523
Reviewed by Yusuke Suzuki.

Right now when we tier up a function to BBQ/OMG we set the JSEntrypointCallee's entrypoint
to the prologue of the newly compiled code. This is wrong because we share JSEntrypointCallee
between BoundsChecking and Signaling instances. This means if tier up a Signaling instance to
we could inadvertently call the Signaling code from a BoundsChecking instance.

To fix this the interpreted/shared JIT JSToWasm entrypoints now do the same logic as the
inline cache entrypoint and load the callee's entrypoint from the indirect location. This
means we need to keep the JS WebAssemblyFunction until we're basically ready to call the
into wasm.

Finally, since there's basically no scratches left I added a BoxedNativeCalleePtrTraits
so the Ref<> of the target Callee can be saved as a boxed pointer rather than as a raw
pointer. So now we no longer need to box the callee at runtime.

Merge-back additional changes:

Use CalleeBits when passing around the boxed callee rather than EncodedJSValue. This
made little sense as Wasm::Callee is not a GC pointer. When I first wrote this patch
we were still using a `uintptr_t` for the boxed callee.

Originally-landed-as: 289651.47@safari-7621-branch (8f74b8980eba). rdar://148059830
Canonical link: https://commits.webkit.org/294376@main