run-jsc-stress-tests --treat-failing-as-flaky should also limit total runtime https://bugs.webkit.org/show_bug.cgi?id=290906
Reviewed by Justin Michaud.

When we treat failing tests as flaky we retry them up to a
user-specified number of iterations, until it's clear whether they can
make the pass percentage. This works well for tests that finish quickly
but gets us in pathological behavior when a test hangs. We then run it
until it hits the timeout multiple times, needlessly prolonguing the
time it tests to run tests in the buildbot.

Add a cap on the number of seconds we can spend retrying and refuse to
start a new iteration if it has been exceeded.

* Tools/Scripts/run-javascriptcore-tests:
* Tools/Scripts/run-jsc-stress-tests:
* Tools/Scripts/webkitruby/jsc-stress-test/executor.rb:

Canonical link: https://commits.webkit.org/294423@main