BBC - Viewer size is not correct when exiting full screen https://bugs.webkit.org/show_bug.cgi?id=292389 rdar://149315208
Reviewed by Jer Noble.

This patch deletes a code block that breaks our enter/exit fullscreen animation
on sites such as BBC and YouTube that are resizing the video during the animation.
The code block returns the frame of the enlarged element which is incorrect; we want
the frame before the element has changed size to be used as the value of
WKFullscreenWindowController's m_initialFrame.

This block attempts to handle the case where the video element has a css
transformation applied to it, such as a rotation. However it seems that situation is
rare, and the code is mishandling this more common case, so deleting it will fix more
cases than break.

* Source/WebKit/WebProcess/FullScreen/WebFullScreenManager.cpp:
(WebKit::screenRectOfContents):

Canonical link: https://commits.webkit.org/294447@main