[GTK][WPE] Enable Cookie Store API by default https://bugs.webkit.org/show_bug.cgi?id=292416
Reviewed by Rupin Mittal and Michael Catanzaro.

Cookie Store API <https://wicg.github.io/cookie-store>
is a Web spec that is currently supported by the three
main Web engines.

Before 289626@main it was enabled by default for the
Apple ports, and after that it depends on a build-time
conditional macro that this patch defaults to enable
for GTK and WPE.

I tested to enable it by default for the GTK and WPE
ports and I have ran the WPT tests (not the imported
ones here in webkit, but the real WPT tests) and most
of the tests pass now.

There are no changes to the layout test expectations
with this patch because WTR runs by default with all
the run-time settings enabled, so layout tests were
already running with this enabled.

* Source/WTF/wtf/PlatformEnableGlib.h:

Canonical link: https://commits.webkit.org/294448@main