[UnifiedPDF] PDFPageNumberIndicator cannot use UIVisualEffectView in all configurations https://bugs.webkit.org/show_bug.cgi?id=292454 rdar://149981713
Reviewed by Aditya Keerthi and Richard Robinson.

UIVisualEffectView is not appropriate for the backdrop of the page
number indicator under certain platform configurations. In such cases,
this patch introduces a path to fallback to a generic UIView that has
its background appropriately configured to not regress the indicator's
functionality.

* Source/WebKit/Configurations/WebKit.xcconfig:
* Source/WebKit/DerivedSources-output.xcfilelist:
* Source/WebKit/DerivedSources.make:
* Source/WebKit/Modules/Internal/WebKitInternal.h:
* Source/WebKit/UIProcess/PDF/WKPDFPageNumberIndicator.h:
* Source/WebKit/UIProcess/PDF/WKPDFPageNumberIndicator.mm:
(-[WKPDFPageNumberIndicator initWithFrame:view:pageCount:]):

Note that when the backdrop is not a UIVisualEffectView, we must take
care to add the page number label directly to the backdrop view, and not
to the `contentView` under the backdrop.

(-[WKPDFPageNumberIndicator backdropView]):
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/294462@main