REGRESSION(294462@main): UI process may crash with 'NSInvalidArgumentException': -[WKPDFPageNumberIndicator canUseVisualEffectViewForBackdrop] :: unrecognized selector https://bugs.webkit.org/show_bug.cgi?id=292531 rdar://150637116
Reviewed by David Kilzer.

Under certain configurations, the `canUserVisualEffectViewForBackdrop`
getter is not implemented, even though we still do have support for the
page number indicator at large. This results in the UI process crashing
with an `unrecognized selector` exception during indicator bring-up.

This patch mitigates said crash by ensuring there is a definition for
the property getter in question before calling into it. We default to
using an UIVisualEffectView for the indicator's backdrop in case the
getter is not implemented.

* Source/WebKit/UIProcess/PDF/WKPDFPageNumberIndicator.mm:
(-[WKPDFPageNumberIndicator initWithFrame:view:pageCount:]):

Canonical link: https://commits.webkit.org/294516@main