AX: Make AXTextRuns a shared_ptr in the AXPropertyValueVariant https://bugs.webkit.org/show_bug.cgi?id=292575 rdar://problem/150721813
Reviewed by Tyler Wilcock.

The size of a variant is relative to the largest type within it, and in the
AXPropertyValueVariant, that was the AXTextRuns, which was 32 bytes (in part
due to memory alignment). This made the overall variant 48 bytes after
padding.

To reduce this size, we can instead store shared_ptr pointing to the text runs,
which only uses 16 bytes. This makes the overall variant 32 bytes after
padding.

While making this change, I also moved the type we used to store the marked
text marker range to be a shared_ptr, since this was the next-largest type at
24 bytes.

This saves 33% of memory per-property that we store in the properties vector.

* Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.cpp:
(WebCore::AXIsolatedObject::initializeProperties):
(WebCore::AXIsolatedObject::setProperty):
(WebCore::AXIsolatedObject::textRuns const):
(WebCore::AXIsolatedObject::textInputMarkedTextMarkerRange const):
* Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.cpp:
(WebCore::AXIsolatedTree::updateNodeProperties):
* Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.h:

Canonical link: https://commits.webkit.org/294565@main