Internal build fix for "[Xcode] Some targets accessing /System/Library/PrivateFrameworks through non-system include paths" https://bugs.webkit.org/show_bug.cgi?id=292451 rdar://150539568
Unreviewed. Previous changes moved WK_QUOTED_OVERRIDE_FRAMEWORKS_DIR to
SYSTEM_FRAMEWORK_SEARCH_PATHS. It turns out that this build setting
is always a path inside the DSTROOT, and does not include the SDK_DIR,
but Xcode silently converts entries in FRAMEWORK_SEARCH_PATHS to SDK
paths when such a path exists.

We should rework WK_QUOTED_OVERRIDE_FRAMEWORKS_DIR entirely, but in the
meantime, put it back in FRAMEWORK_SEARCH_PATHS to fix staged framework
builds. The cause of the original bug was NORMAL_UMBRELLA_FRAMEWORKS_DIR
not being treated as a system path, so this should not affect the fix.

* Source/WebInspectorUI/Configurations/Base.xcconfig:
* Tools/DumpRenderTree/mac/Configurations/Base.xcconfig:
* Tools/WebKitTestRunner/Configurations/Base.xcconfig:

Canonical link: https://commits.webkit.org/294648@main