[GTK][WPE] No damage produced when layer's opacity is animated https://bugs.webkit.org/show_bug.cgi?id=292768
Reviewed by Carlos Garcia Campos.

This change fixes the lack of damage when damage propagation is enabled
and when layer's opacity is animated fast.

The problem was due to the fact that opacity animation has a custom
handling code in the TextureMapperLayer and it was not taken into
account when inferring damage.

* LayoutTests/platform/glib/damage/animations-opacity-expected.txt: Added.
* LayoutTests/platform/glib/damage/animations-opacity.html: Added.
* Source/WebCore/platform/graphics/texmap/TextureMapperLayer.cpp:
(WebCore::TextureMapperLayer::syncAnimations):

Canonical link: https://commits.webkit.org/294705@main