[Flex] Refactor computeNextFlexLine to return a struct https://bugs.webkit.org/show_bug.cgi?id=292748 rdar://150969447
Reviewed by Alan Baradlay.

This function signature has many arguments that are passed in by
reference to use to write data out. We can use a struct instead which
improves the readability of the codebase by decluttering it a bit and
the maintainability by encapsulating the data together.

I named this struct "FlexingLineData," because it contains the flex
items which are placed on the line along with other pieces of data which
are used as input to flexing the content on the line.

* Source/WebCore/rendering/FlexLineBuilder.cpp:
(WebCore::FlexLineBuilder::computeNextFlexLine):
* Source/WebCore/rendering/FlexLineBuilder.h:
* Source/WebCore/rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::performFlexLayout):

Canonical link: https://commits.webkit.org/294706@main