Remove Quirk needsIPadMiniUserAgent for huya.com https://bugs.webkit.org/show_bug.cgi?id=284508 rdar://141335024
Reviewed by Brent Fulgham.

The volume control slider is not working properly on huya.com when
displaying the desktop version of the site on iPad, but this is no
different than many websites currently on iPad. We need to find a
more systematic way to make the volume control slider work properly
on all websites. Knowing that, this issue has no specific purpose
and the code targetting specifically huya.com is not necessary.
We remove the Quirk which was giving huya.com an iPad Mini
User Agent instead of the default desktop user agent.

* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::needsIPadMiniUserAgent):

Canonical link: https://commits.webkit.org/294713@main