Lexer errors should be cleared before parsing class default initialization expressions https://bugs.webkit.org/show_bug.cgi?id=292745 rdar://149546540
Reviewed by Yusuke Suzuki.

Currently, when we have a default initialization expression, we re-parse the function
in strict mode. To do this, we create a new parser on the entire source code, and set
the offset of the parser to the selected expression. However, since the parser passes
through all of the code in the script, the lexer may produce an error on code that we
do not parse for the expressions. This can lead to debug assertions failing due to an
incorrectly set error bit.

* Source/JavaScriptCore/parser/Parser.cpp:
(JSC::Parser<LexerType>::parseClassFieldInitializerSourceElements):

Canonical link: https://commits.webkit.org/294718@main