[Quirk] Allow users to upload HEIC images on zillow.com https://bugs.webkit.org/show_bug.cgi?id=292041 rdar://79872092
Reviewed by Brent Fulgham.

When users of the zillow websites want to upload an HEIC images,
they get a message saying:

    HEIC format is not supported on web. Please download
    the Zillow Rental Manager App to upload, or convert
    the files to .png, .jpg, .jpeg, or .gif to upload.

WebKit added the possibility to automatically convert the HEIC image
format to one that the site accepts.
https://github.com/WebKit/WebKit/commit/1fd3fe3f407ad3cbc9fde83b03a9a0641e6d3e58

The current input on the zillow website is

    <input accept="image/*" multiple="" type="file">

To benefit from WebKit capability to auto-convert the file
to the format of website choice, it is necessary to specify
in the accept attribute, the accepted format, for example:

    accept="image/jpeg, image/png, image/gif"

This quirk is forcing for zillow image/jpeg as a conversion target.

* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldTranscodeHeicImagesForURL):
* Source/WebCore/page/Quirks.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::didChooseFilesForOpenPanelWithImageTranscoding):

Canonical link: https://commits.webkit.org/294726@main